+++ docs/generate.py
@@ -4,7 +4,7 @@ import importlib
 import os
 from ruamel.yaml import YAML
 
-from openbb_terminal.api import functions
+from openbb_terminal.sdk import functions
 
 # NOTE: The main.yml and documentation _index.md files are automaticallty overridden
 # every time this is ran. Folder level _index.md files are NOT overridden after creation
