+++ tests/openbb_terminal/api/test_api.py
@@ -6,7 +6,6 @@
 
 
 def test_load_api():
-    # pylint: import-outside-toplevel
-    from openbb_terminal import api
+    from openbb_terminal import sdk
 
-    print(api)
+    print(sdk)
